refactor: rename categories hook utility
